refactor: ディレクトリ構成を適切にした
