+++ src/models/bot.py
