+++ src/services/message_handler.py
@@ -1,11 +1,11 @@
 import re
-from services.bot import Bot
-from services.reaction_handler import ReactionHandler
-from services.openai_client import OpenAIClient
-from services.config_service import ConfigService
-from services.custom_random import random_true_with_probability
-from role_mention_checker import check_role_mention
-from services.prompt_loader import get_prompt
+from src.models.bot import Bot
+from src.services.reaction_handler import ReactionHandler
+from src.services.openai_client import OpenAIClient
+from src.services.config_service import ConfigService
+from src.utils.random_utils import random_true_with_probability
+from src.utils.role_mention_checker import check_role_mention
+from src.services.prompt_loader import get_prompt
 
 class MessageHandler:
     def __init__(
